made FS_WriteFile's len parameter size_t to fix some warnings
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5491 d7cf8633-e32d-0410-b094-e92efae38249

+++ fs.h
@@ -70,7 +70,7 @@ fssearch_t *FS_Search(const char *pattern, int caseinsensitive, int quiet);
 void FS_FreeSearch(fssearch_t *search);
 
 qbyte *FS_LoadFile (const char *path, mempool_t *pool, qboolean quiet);
-qboolean FS_WriteFile (const char *filename, void *data, int len);
+qboolean FS_WriteFile (const char *filename, void *data, size_t len);
 
 
 // ------ Other functions ------ //
